% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/things.R
\name{get_list}
\alias{get_list}
\title{Retrieves the contents of a named list from list.design}
\usage{
get_list(stuff, json = FALSE)
}
\arguments{
\item{stuff}{name of the list you'd like to retrieve}

\item{json}{Set as TRUE if you want unparsed JSON}
}
\value{
\code{data.frame}
}
\description{
Retrieves the contents of a named list from list.design
}
\examples{
library(magrittr)
stuff() \%>\% .[[1]] \%>\% get_list()
}
